requirements.txt
# Backend Core Dependencies
fastapi>=0.109.1
uvicorn[standard]>=0.27.0
pydantic>=2.6.1
python-dotenv>=1.0.0

# LangChain and Related
langchain>=0.1.0
langchain-core>=0.1.10
langchain-community>=0.0.10
langchain-google-community>=0.0.10

# Vector Store
faiss-cpu>=1.7.4
chromadb>=0.4.22

# Embedding and ML Models
sentence-transformers>=2.3.1
transformers>=4.36.2
torch>=2.2.0
accelerate>=0.26.1
sentencepiece>=0.1.99  # Added sentencepiece

# Google Search API
google-api-python-client>=2.118.0
google-auth>=2.27.0

# Utilities
numpy>=1.24.3
pandas>=2.1.4
tqdm>=4.66.1
pyzmq>=24.0.0
matplotlib>=3.4.0
cloudpickle>=3.0.0
gitpython>=3.1.9
pyarrow>=4.0.0,<18.0.0
antlr4-python3-runtime==4.9.1
pyyaml>=6.0.1  # Added pyyaml

# API and Web
httpx>=0.27.0
requests>=2.31.0

# Development Tools
black>=23.12.1
isort>=5.13.2
pytest>=7.4.4
python-multipart>=0.0.6
